♻️ refactor: chat관련 페이지 metadata 추가
